+++ boofuzz/monitors/pedrpc.py
@@ -1,3 +1,4 @@
+import errno
 import pickle
 import select
 import socket
@@ -45,6 +46,7 @@ class Client(object):
 
         # connect to the server, timeout on failure.
         self.__server_sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+        self.__server_sock.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
         self.__server_sock.settimeout(3.0)
         try:
             self.__server_sock.connect((self.__host, self.__port))
@@ -114,10 +116,10 @@ class Client(object):
         # connect to the PED-RPC server.
         self.__connect()
 
-        uuid = self.__pickle_recv()
-        if uuid != self.known_server:
-            self.on_new_server(uuid)
-            self.known_server = uuid
+        server_uuid = self.__pickle_recv()
+        if server_uuid != self.known_server:
+            self.on_new_server(server_uuid)
+            self.known_server = server_uuid
 
         # transmit the method name and arguments.
         self.__pickle_send((method_name, (args, kwargs)))
@@ -213,6 +215,7 @@ class Server(object):
         self.__dbg_flag = False
         self.__client_sock = None
         self.__client_address = None
+        self.__running = True
 
         # This is a bad solution for a problem that should not even exist in the first place.
         # The Problem is that the client disconnects after each RPC call,
@@ -235,10 +238,11 @@ class Server(object):
         try:
             # create a socket and bind to the specified port.
             self.__server = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+            self.__server.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
             self.__server.settimeout(None)
             self.__server.bind((host, port))
             self.__server.listen(1)
-        except Exception:
+        except socket.error:
             sys.stderr.write("unable to bind to %s:%d\n" % (host, port))
             sys.exit(1)
 
@@ -249,8 +253,14 @@ class Server(object):
 
         if self.__client_sock is not None:
             self.__debug("closing client socket")
+            try:
+                self.__client_sock.shutdown(socket.SHUT_RDWR)
+            except socket.error as e:
+                if e.errno == errno.ENOTCONN:
+                    pass
+                else:
+                    raise
             self.__client_sock.close()
-            self.__client_sock = None
 
     def __debug(self, msg):
         if self.__dbg_flag:
@@ -306,12 +316,12 @@ class Server(object):
     def serve_forever(self):
         self.__debug("serving up a storm")
 
-        while 1:
+        while self.__running:
             # close any pre-existing socket.
             self.__disconnect()
 
             # accept a client connection.
-            while True:
+            while self.__running:
                 readable, writeable, errored = select.select([self.__server], [], [], 0.1)
                 if len(readable) > 0:
                     assert readable[0] == self.__server
@@ -322,7 +332,7 @@ class Server(object):
 
             self.__pickle_send(self.__instance)
 
-            # recieve the method name and arguments, continue on socket disconnect.
+            # receive the method name and arguments, continue on socket disconnect.
             try:
                 (method_name, (args, kwargs)) = self.__pickle_recv()
                 self.__debug("%s(args=%s, kwargs=%s)" % (method_name, args, kwargs))
@@ -341,3 +351,15 @@ class Server(object):
                 self.__pickle_send(ret)
             except Exception:
                 continue
+
+    def stop(self):
+        self.__running = False
+        self.__disconnect()
+        try:
+            self.__server.shutdown(socket.SHUT_RDWR)
+        except socket.error as e:
+            if e.errno == errno.ENOTCONN:
+                pass
+            else:
+                raise
+        self.__server.close()